backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.2
Pillow==10.4.0
openai==1.43.0
